Set the graphics buffer and add noise raindrop animation
In order to animate the team work, I need to draw all the elements as a background and update the background frequently in the draw function.
I used the createGraphics function of p5.js to draw all the elements in the graphics buffer outside the screen and then draw it as image in Draw.
I need to modify each function to accept buffer.
In addition, I added a noise raindrop animation to simulate apple trees in the rain.